Windows and Linux define address families as unsiged short, yet the inet_ntop function takes it as int

+++ kissnet.hpp
@@ -153,7 +153,7 @@ inline const char* inet_ntop(int af, const void* src, char* dst, socklen_t size)
 	} addr;
 	int res;
 	memset(&addr, 0, sizeof(addr));
-	addr.sa.sa_family = af;
+	addr.sa.sa_family = (unsigned short) af;
 	if (af == AF_INET6)
 	{
 		memcpy(&addr.sai6.sin6_addr, src, sizeof(addr.sai6.sin6_addr));
